apply changes after moving
